pvt_table_sync/src/requirements.txt
aiofiles==23.1.0
anyio==3.6.2
APScheduler==3.10.1
async-timeout==4.0.2
backcall==0.2.0
debugpy==1.6.6
email-validator==1.3.1
exceptiongroup==1.1.1
fastapi==0.93.0
mysql-connector-python==8.0.32
pydantic==1.10.6
psycopg2-binary==2.9.5
python-dotenv==1.0.0
redis==4.5.1
requests==2.28.2
SQLAlchemy==2.0.6
uvicorn==0.20.0
watchfiles==0.18.1
gspread==6.1.0
openpyxl==3.1.4
pandas==2.2.2
